requirements.txt
streamlit==1.31.1
pillow==8.4.0
opencv-python==4.5.3.56
numpy==1.21.2
click==7.1.2
openai==0.28.0
python-dotenv==1.0.1